[ALPS07977079] use dma_buf api
use dma_buf api

MTK-Commit-Id: 6d48a23791cd79862e2a08d3a8be495d530b6d4b

CR-Id: ALPS07977079
Feature: Video Driver

+++ drivers/misc/mediatek/vcp/rv/vcp_helper.h
@@ -144,7 +144,7 @@ struct vcp_reserve_mblock {
 	enum vcp_reserve_mem_id_t num;
 	u64 start_phys;
 	u64 start_virt;
-	u64 size;
+	u32 size;
 };
 
 struct vcp_region_info_st {
